requirements-gpu.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.3

# Redis
redis==5.0.1
aioredis==2.0.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ML/AI (GPU enabled)
transformers==4.35.2
torch==2.1.1+cu118
accelerate==0.24.1
peft==0.6.2
trl==0.7.4
datasets==2.15.0
tokenizers==0.15.0
sentencepiece==0.1.99
bitsandbytes==0.41.2

# Training & Evaluation
tensorboard==2.15.1
wandb==0.16.0
evaluate==0.4.1
rouge-score==0.1.2
sacrebleu==2.3.1

# LLM APIs
openai==1.3.5
anthropic==0.7.0
google-generativeai==0.3.0

# Monitoring
mlflow==2.8.1
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Task Queue
celery==5.3.4
flower==2.0.1

# Utils
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
Jinja2==3.1.2
loguru==0.7.2
psutil==5.9.6

# Quality Filtering
scikit-learn==1.3.2
nltk==3.8.1
langdetect==1.0.9
numpy==1.24.4
tiktoken==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Serving (GPU enabled)
ray==2.8.0
vllm==0.2.5
aiohttp==3.9.1 